ENR-2646: Change NoOverlap validator instead of creating ChildrenOverlapValidator
