+++ lib/acts_as_span/span_instance.rb
@@ -1,6 +1,5 @@
 require 'acts_as_span/span_instance/validations'
 require 'acts_as_span/span_instance/status'
-require 'acts_as_span/span_instance/date_range'
 
 require 'active_support/core_ext/module/delegation'
 
@@ -46,13 +45,17 @@ module ActsAsSpan
     # @param other [ActsAsSpan::SpanInstance]
     # @return [Boolean]
     def overlap?(other)
-      date_range.overlap?(other.date_range)
+      other.actual_start_date <= actual_end_date && actual_start_date <= other.actual_end_date
     end
 
     protected
 
-    def date_range
-      DateRange.new(start_date: start_date, end_date: end_date)
+    def actual_end_date
+      end_date || Date::Infinity.new
+    end
+
+    def actual_start_date
+      start_date || Date.current
     end
   end
 end
