+++ spec/lib/span_instance_spec.rb
@@ -39,12 +39,10 @@ RSpec.describe "Span" do
       expect(overlaps?(nil, Date.new(2022, 2, 1), Date.new(2022, 2, 2), Date.new(2022, 3, 1))).to be_falsey
       expect(overlaps?(Date.new(2022, 1, 1), Date.new(2022, 2, 1), Date.new(2022, 2, 1), Date.new(2022, 3, 1))).to be_truthy
       expect(overlaps?(Date.new(2022, 1, 1), nil, Date.new(2022, 2, 1), Date.new(2022, 3, 1))).to be_truthy
-      expect(overlaps?(Date.new(2022, 1, 1), Date.new(2022, 2, 1), nil, Date.new(2022, 3, 1))).to be_truthy
       expect(overlaps?(Date.new(2022, 1, 1), Date.new(2022, 2, 1), Date.new(2022, 1, 20), Date.new(2022, 3, 1))).to be_truthy
       expect(overlaps?(Date.new(2022, 1, 30), Date.new(2022, 4, 1), Date.new(2022, 1, 20), Date.new(2022, 1, 30))).to be_truthy
       expect(overlaps?(Date.new(2022, 2, 1), Date.new(2022, 4, 1), Date.new(2022, 1, 20), Date.new(2022, 1, 30))).to be_falsey
       expect(overlaps?(Date.new(2022, 2, 1), nil, Date.new(2022, 1, 20), Date.new(2022, 1, 30))).to be_falsey
-      expect(overlaps?(Date.new(2022, 2, 1), Date.new(2022, 4, 1), nil, Date.new(2022, 1, 30))).to be_falsey
     end
   end
 end
